chore: renamed old component with `_old` postfix, added switcher by flag condition (M2-6841)
