+++ src/modules/Builder/components/index.ts
@@ -8,4 +8,4 @@ export * from './SaveChangesPopup';
 export * from './DndDroppable';
 export * from './InsertItem';
 export * from './ToggleItemContainer';
-export * from './ConditionRow';
+export * from './ConditionRow_old';
