+++ src/modules/Builder/features/ActivityItemsFlow_old/ItemFlow/ItemFlowContent/StaticConditionRow/StaticConditionRow.tsx
