+++ src/modules/Builder/features/SaveAndPublish/SaveAndPublish.hooks.ts
@@ -37,6 +37,7 @@ import {
 } from 'modules/Builder/utils/getEntityReportFields';
 import { banners } from 'shared/state/Banners';
 import { ErrorResponseType, LocationState, LocationStateKeys } from 'shared/types';
+import { useFeatureFlags } from 'shared/hooks/useFeatureFlags';
 
 import {
   getActivityItems,
@@ -131,7 +132,8 @@ export const useCheckIfHasAtLeastOneItem = () => {
 
 export const useCheckIfHasEmptyRequiredFields = () => {
   const { getValues } = useCustomFormContext();
-  const appletSchema = AppletSchema();
+  const { featureFlags } = useFeatureFlags();
+  const appletSchema = AppletSchema(featureFlags.enableItemFlowExtendedItems);
 
   return async () => {
     const body = getValues();
@@ -148,7 +150,8 @@ export const useCheckIfHasEmptyRequiredFields = () => {
 
 export const useCheckIfHasErrorsInFields = () => {
   const { getValues } = useCustomFormContext();
-  const appletSchema = AppletSchema();
+  const { featureFlags } = useFeatureFlags();
+  const appletSchema = AppletSchema(featureFlags.enableItemFlowExtendedItems);
 
   return async () => {
     const body = getValues();
