+++ src/modules/Builder/routes/routes.const.ts
@@ -9,7 +9,7 @@ const ActivityAbout = lazy(() => import('modules/Builder/features/ActivityAbout'
 const ActivityFlowAbout = lazy(() => import('modules/Builder/features/ActivityFlowAbout'));
 const ActivityFlow = lazy(() => import('modules/Builder/features/ActivityFlow'));
 const ActivityFlowBuilder = lazy(() => import('modules/Builder/features/ActivityFlowBuilder'));
-const ActivityItemsFlow = lazy(() => import('modules/Builder/features/ActivityItemsFlow'));
+const ActivityItemsFlowOld = lazy(() => import('modules/Builder/features/ActivityItemsFlow_old'));
 const Flanker = lazy(() => import('modules/Builder/features/PerformanceTasks/Flanker'));
 const GyroscopeAndTouch = lazy(
   () => import('modules/Builder/features/PerformanceTasks/GyroscopeAndTouch'),
@@ -21,9 +21,9 @@ export const appletRoutes = [
   { path: Path.ActivityFlow, Component: ActivityFlow },
 ];
 
-export const appletActivityRoutes = [
+export const appletActivityRoutes = (enableItemFlowExtendedItems: boolean) => [
   { path: Path.About, Component: ActivityAbout },
-  { path: Path.ItemsFlow, Component: ActivityItemsFlow },
+  { path: Path.ItemsFlow, Component: ActivityItemsFlowOld },
 ];
 
 export const appletActivityFlowRoutes = [
